chore(workflow): add commitlint and commit template
Adds commitlint to the repository to validate commit messages. Also adds
a script to add a default git commit template. This can usually be done
with `git config commit.template <FILE>`, however, this would require
all contributors to run the command to get this functionality. The
script is run with the git hooks installed by husky, and so is set up
when husky is installed (with npm install).